\name{nnTensor-package}
\alias{nnTensor-package}
\alias{nnTensor}
\docType{package}
\title{
\packageTitle{nnTensor}
}
\description{
\packageDescription{nnTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nnTensor}
\packageIndices{nnTensor}
}
\author{
\packageAuthor{nnTensor}

Maintainer: \packageMaintainer{nnTensor}
}
\references{
Andrzej CICHOCK, et. al., (2009). Nonnegative Matrix and Tensor Factorizations. \emph{John Wiley & Sons, Ltd}

Keigo Kimura, (2017). A Study on Efficient Algorithms for Nonnegative Matrix/Tensor Factorization. \emph{Hokkaido University Collection of Scholarly and Academic Papers}

Andrzej CICHOCKI et. al., (2007). Non-negative Tensor Factorization using Alpha and Beta Divergence. \emph{IEEE ICASSP 2007}

Anh Huy PHAN et. al., (2008). Multi-way Nonnegative Tensor Factorization Using Fast Hierarchical Alternating Least Squares Algorithm (HALS). \emph{NOLTA2008}

Andrzej CICHOCKI et. al., (2008). Fast Local Algorithms for Large Scale Nonnegative Matrix and Tensor Factorizations. \emph{IEICE Transactions on Fundamentals of Electronics, Communications and Computer Sciences}

Yong-Deok Kim et. al., (2007). Nonnegative Tucker Decomposition. \emph{IEEE Conference on Computer Vision and Pattern Recognition}

Yong-Deok Kim et. al., (2008). Nonneegative Tucker Decomposition With Alpha-Divergence. \emph{IEEE International Conference on Acoustics, Speech and Signal Processing}

Anh Huy Phan, (2008). Fast and efficient algorithms for nonnegative Tucker decomposition. \emph{Advances in Neural Networks - ISNN2008}

Anh Hyu Phan et. al. (2011). Extended HALS algorithm for nonnegative Tucker decomposition and its applications for multiway analysis and classification. \emph{Neurocomputing}

Jean-Philippe Brunet. et. al., (2004). Metagenes and molecular pattern discovery using matrix factorization. \emph{PNAS}

Xiaoxu Han. (2007). CANCER MOLECULAR PATTERN DISCOVERY BY SUBSPACE CONSENSUS KERNEL CLASSIFICATION

Attila Frigyesi. et. al., (2008). Non-Negative Matrix Factorization for the Analysis of Complex Gene Expression Data: Identification of Clinically Relevant Tumor Subtypes. \emph{Cancer Informatics}

Haesun Park. et. al., (2019). Lecture 3: Nonnegative Matrix Factorization: Algorithms and Applications. \emph{SIAM Gene Golub Summer School, Aussois France, June 18, 2019}

Chunxuan Shao. et. al., (2017). Robust classification of single-cell transcriptome data by nonnegative matrix factorization. \emph{Bioinformatics}

Paul Fogel (2013). Permuted NMF: A Simple Algorithm Intended to Minimize the Volume of the Score Matrix

Philip M. Kim. et. al., (2003). Subsystem Identification Through Dimensionality Reduction of Large-Scale Gene Expression Data. \emph{Genome Research}

Lucie N. Hutchins. et. al., (2008). Position-dependent motif characterization using non-negative matrix factorization. \emph{Bioinformatics}

Patrik O. Hoyer (2004). Non-negative Matrix Factorization with Sparseness Constraints. \emph{Journal of Machine Learning 5}
}

\keyword{ package }
\seealso{\code{\link{toyModel}},\code{\link{NMF}},\code{\link{NTF}},\code{\link{NTD}},\code{\link{recTensor}},\code{\link{plotTensor3D}}}
\examples{
ls("package:nnTensor")
}
